230205_AlbatrossReminder_3.0.0(D) 메인 화면 및 수정 화면 전반적 수정

+++ app/src/main/java/net/ienlab/sogangassist/room/LMSEntity.kt
@@ -30,9 +30,9 @@ class LMSEntity(
         const val TYPE_EXAM = 5
     }
 
-    @PrimaryKey(autoGenerate = true) var id: Int? = null
+    @PrimaryKey(autoGenerate = true) var id: Long? = null
 
-    fun setDataId(id: Int) {
+    fun setDataId(id: Long) {
         this.id = id
     }
